# Use RWLocks for counter measurements