@dashpole 2025-09-24
Blocked on:

* https://github.com/open-telemetry/opentelemetry-go/pull/7407
* https://github.com/open-telemetry/opentelemetry-go/pull/7406
* https://github.com/open-telemetry/opentelemetry-go/pull/7379