@dashpole 2025-09-30
Closing in favor of https://github.com/open-telemetry/opentelemetry-go/pull/7427